implemnted group search with lookup

+++ src/vectorDB/push_qdrant.py
@@ -36,9 +36,10 @@ class QdrantManager:
                  model_name: str = qdrant_config["model_name"],
                  docker_timeout: float = qdrant_config["docker_timeout"],
                  watch_tmp_timeout: float = qdrant_config["watch_tmp_timeout"],
-                 max_threads: int = qdrant_config["max_threads"],):
+                 max_threads: int = qdrant_config["max_threads"],
+                 clean_after_push: bool = True):
 
-        self.volume_path = config.VECTORDB_PATH
+        self.volume_path = config.VECTORDB_PATH / "Qdrant"
         self.tmp_path = config.VECTORDB_PATH / "tmp"
         self.container_name = config.PROJECT_NAME + "-Qdrant"
         self.collection_name = model_name.split('/')[-1]
@@ -49,8 +50,10 @@ class QdrantManager:
         self.chunk_fields = {'paperId', 'fieldsOfStudy', 'venue', 'authorName',
                              'citationCount', 'influentialCitationCount',
                              'publicationDate', 'publicationYear', 'text'}
-        self.id_fields = {'paperId', 'openAccessPdf',
-                          'title', 'abstract', 'full_text'}
+        self.id_fields = {'paperId', 'fieldsOfStudy', 'venue', 'authorName',
+                          'citationCount', 'influentialCitationCount',
+                          'publicationDate', 'publicationYear', 'openAccessPdf',
+                          'title', 'abstract', 'main_text'}
 
         hidden_size = AutoConfig.from_pretrained(self.model_name).hidden_size
         self.vector_size = hidden_size
@@ -61,6 +64,7 @@ class QdrantManager:
         self.docker_timeout = docker_timeout
         self.watch_tmp_timeout = watch_tmp_timeout
         self.max_threads = max_threads
+        self.clean_after_push = clean_after_push
 
         self.start_qdrant_container()
         self.client = AsyncQdrantClient(host=self.host, port=self.port)
@@ -163,18 +167,22 @@ class QdrantManager:
             data = json.load(f)
         embeddings = data['embeddings']
         payloads = data['payloads']
+        ref_embedding = data['ref_embedding']
         chunk_points = [PointStruct(id=str(uuid.uuid5(namespace=uuid.NAMESPACE_DNS, name=f"{paperId}_{i}")),
                                     vector=emb, payload={field: payloads[i][field] for field in self.chunk_fields})
                         for i, emb in enumerate(embeddings)]
 
         await self.client.upsert(collection_name=self.collection_name_chunks, points=chunk_points)
-
-        id_points = [PointStruct(id=str(uuid.uuid5(namespace=uuid.NAMESPACE_DNS, name=f"{paperId}")),
-                                 vector=[0], payload={field: payloads[0][field] for field in self.id_fields})]
-
+        try:
+            id_points = [PointStruct(id=str(uuid.uuid5(namespace=uuid.NAMESPACE_DNS, name=f"{paperId}")),
+                                     vector=ref_embedding, payload={field: payloads[0][field] for field in self.id_fields})]
+        except Exception as e:
+            print(paperId)
+            raise e
         await self.client.upsert(collection_name=self.collection_name_ids, points=id_points)
 
-        os.remove(file_path)
+        if self.clean_after_push:
+            os.remove(file_path)
 
     async def check_collection(self):
         if not await self.client.collection_exists(self.collection_name_ids):
@@ -207,23 +215,27 @@ class QdrantManager:
 
         await self.check_collection()
 
-        paper_ids = []
-        while paper_ids or not done_event.is_set():
+        new_ids = self.get_embedding_ids()
+        pushed_ids = []
+        while new_ids or not done_event.is_set():
             paper_ids = self.get_embedding_ids()
-            if paper_ids:
-                await self.delete_papers(paper_ids)
+            new_ids = [id for id in paper_ids if id not in pushed_ids]
+            if new_ids:
+                await self.delete_papers(new_ids)
 
-                upload_tasks = [self.upload_paper(id) for id in paper_ids]
+                upload_tasks = [self.upload_paper(id) for id in new_ids]
                 with tqdm(asyncio.as_completed(upload_tasks),
                           total=len(upload_tasks),
                           desc='Uploading') as pbar:
                     for f in pbar:
                         await f
+
+            pushed_ids = paper_ids
             time.sleep(5)
 
         self.stop_qdrant_container()
 
-        if os.path.isdir(self.tmp_path):
+        if os.path.isdir(self.tmp_path) and self.clean_after_push:
             os.removedirs(self.tmp_path)
 
     async def query(self,
@@ -250,9 +262,9 @@ class QdrantManager:
         query_filter = Filter(
             must=must_conditions) if must_conditions else None
 
-        results = await self.client.search(
+        results = await self.client.query_points_groups(
             collection_name=self.collection_name_chunks,
-            query_vector=query_vector,
+            query=query_vector,
             query_filter=query_filter,
             group_by="paperId",
             limit=topK,  # Max amount of groups
@@ -271,7 +283,7 @@ class QdrantManager:
                                    topK_per_paper: int = 5, **kwargs):
         tasks = [
             self.query(query_vector=vec,
-                       top_k=topK,
+                       topK=topK,
                        topK_per_paper=topK_per_paper,
                        **kwargs)
             for vec in query_vectors
@@ -288,7 +300,7 @@ class QdrantManager:
                                      topK_per_paper: int = 5, **kwargs) -> AsyncGenerator:
         tasks = [
             self.query(query_vector=vec,
-                       top_k=topK,
+                       topK=topK,
                        topK_per_paper=topK_per_paper,
                        **kwargs)
             for vec in query_vectors
